refactor: 리액션 api get 커스텀 훅 분리
