+++ src/components/header/components/EmojiShareWrapper.jsx
@@ -1,4 +1,6 @@
 import { RecipientsReactionsAPI } from 'data/CallAPI';
+import useAllReactions from 'hooks/useAllReactions';
+import useReactions from 'hooks/useReactions';
 import { useEffect, useState } from 'react';
 import { useMediaQuery } from 'react-responsive';
 import { useParams } from 'react-router-dom';
@@ -10,43 +12,14 @@ function EmojiShareWrapper({ setShowToast, isLoading }) {
   const [isOpenReactionList, setIsOpenReactionList] = useState(false);
   const [currentEmoji, setCurrentEmoji] = useState('');
   const isDesktop = useMediaQuery({ minWidth: 1024 });
-  const { postId } = useParams();
 
-  const [reactions, setReactions] = useState([]);
-  const [allReactions, setAllReactions] = useState([]);
-  const [reactionCount, setReactionCount] = useState(0);
-
-  const getReactions = async () => {
-    const limit = 3;
-    try {
-      const response = await RecipientsReactionsAPI('get', postId, null, limit);
-      setReactions(response.results);
-    } catch (error) {
-      console.error(error);
-    }
-  };
+  const { reactions, getReactions } = useReactions();
+  const { allReactions, getAllReactions } = useAllReactions();
 
   useEffect(() => {
     getReactions();
     getAllReactions();
-  }, []);
-
-  const getAllReactions = async (all = false) => {
-    let limit = isDesktop ? 8 : 6;
-    limit = all ? reactionCount : null;
-    try {
-      const response = await RecipientsReactionsAPI('get', postId, null, limit);
-      setAllReactions(response.results);
-      setReactionCount(response.count);
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
-  useEffect(() => {
-    getReactions();
-    getAllReactions();
-  }, [postId, isOpenReactionList, isDesktop, currentEmoji]);
+  }, [isOpenReactionList, isDesktop, currentEmoji]);
 
   return (
     <div className='emoji-share-wrapper'>
